fixed a strange edge case
There is a weird edge case with game id 109993492104521301 where one of the players on the red team completely is missing the esportsPlayerId. This player also does not show up under the professional players. In cases where players are missing their player id, the value of their participant id will be used in its place. This will help if there are ever multiple players that are missing their esports ids on the same team in the same game.

This is not a great fix, but there is nothing that can be done when the data from riot does not exist.