Users can update scoring settings
Users can update the scoring settings on the fantasy leagues that they are the owner of.
The fantasy league id in the request body is ignored and the id from the request path will be used in it's place.
Fantasy league scoring settings can only be updated when the status of the fantasy league is PRE_DRAFT

resolves #195