Start fantasy draft (#291)
Owners of their fantasy league can start the Fantasy Draft by moving the fantasy league into `DRAFT` status.
If the fantasy league has too few or too many members an error is raised. If there are no set available leagues an error is raised. If the fantasy league is not in `PRE_DRAFT` status an error is raised.

resolves #271