Fixed an issue with getting schedule
When getting schedule, it assumed that there was going to be a next page when getting the default schedule. This is no longer true.
Now, when getting the default schedule, it will sent the current as the last, then when we go and get the schedule from the current, we will get the token of the default schedule and set it as the default.