Refactor data scraping from riot (#367)
Refactored where and how the scraping from riots APIs take place.

Before it was highly coupled to the API service that served the fetched data from riot. It has been broken out into it's own scraping node. This allows for in the future to scrape multiple games in parallel using child nodes. This will be important when scraping data such as gold diff, multi-kills, etc.

The riot api requester was also refactored to be more generic, returning the entire response rather than parsing the json inside of the api client. This is a more robust, cleaner and correct way for creating API clients and should have been this way from the start.

Also added new fields scraped data such as which side the a team is on (red or blue) in a given game, as well as the game statics about the entire team.